Refactor: Separate API client and configuration from ContentView
- Created LLMService.swift to handle all OpenRouter API calls
  - Moved LLMClient enum (267 lines) from ContentView to dedicated service
  - Implemented LLMServiceProtocol for dependency injection and testability
  - Preserved all functionality: streaming, temporal context, logging

- Created AppConfiguration.swift to centralize app configuration
  - Manages LLM model selection
  - Handles system prompt loading
  - Centralizes API key management
  - Single source of truth for all configuration

- Updated ConversationManager.swift
  - Uses LLMService via dependency injection
  - Simplified public API: sendMessage(_ text: String)
  - Configuration handled internally via AppConfiguration
  - No longer requires callers to pass model, API key, or system prompt

- Cleaned up ContentView.swift
  - Removed 286 lines of API client code
  - Removed configuration logic (model, systemPrompt, apiKey state)
  - Reduced from 882 to ~590 lines
  - Now focused purely on UI presentation
  - Uses simplified ConversationManager API

Benefits:
- Proper separation of concerns (UI ↔ Business ↔ Data)
- Improved testability (can inject mocks)
- Better maintainability (changes isolated to appropriate layers)
- Consistent architecture (follows existing Services/ pattern)
- Zero behavior changes (pure refactoring)
